+++ src/pokedex/serde.go
@@ -1,0 +1,1 @@
+package pokedex
